chore: add module handlers
